+++ Assessment1/src/App.css
@@ -10,7 +10,6 @@
 }
 @media (min-width: 640px) {
   .app {
-    /* width: 100vw; */
     display: flex;
     flex-direction: column;
     flex-wrap: nowrap;
@@ -20,22 +19,18 @@
     overflow-x: hidden;
   }
 }
-
+/* display changes from flex to grid with media queries*/
 @media (min-width: 641px) {
   .grid-items {
     display: grid;
-
-grid-template-columns: repeat(3, 1fr);
-
-grid-column-gap: 15px;
-grid-row-gap: 20px;
-
+    grid-template-columns: repeat(3, 1fr);
+    grid-column-gap: 15px;
+    grid-row-gap: 20px;
   }
 }
- @media (min-width: 1025px) {
-   
-   .grid-items{
-     padding:20px;
+@media (min-width: 1025px) {
+  .grid-items {
+    padding: 20px;
     grid-template-columns: repeat(4, 1fr);
-   }
- }
+  }
+}
